Remove circular dependency and fix tests
